% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_format.R
\name{hex2dec}
\alias{hex2dec}
\title{Convert between pittag formats}
\usage{
hex2dec(x)
}
\arguments{
\item{x}{character vector}
}
\value{
A character vector of of pittags converted to new format.
}
\description{
Pittag ids are 64-bit numbers. The pittag is usuallt presented as a
text string with either decimal och hexadecimal numbers. These functions
convert bwtween these two formats
}
